requirements.txt
html-table-parser-python3
lxml
openpyxl
pandas
python-crontab
python-dateutil
requests

